Change type of struct xlat.val to uint64_t
Some ioctls have flags fields that are 64-bit.  A 32-bit val means
these flags will never be matched or printed.

* xlat.h: Include <stdint.h>.
(struct xlat): Change type of val to uint64_t.

+++ xlat.h
@@ -1,7 +1,9 @@
 #ifndef STRACE_XLAT_H
 
+# include <stdint.h>
+
 struct xlat {
-	unsigned int val;
+	uint64_t val;
 	const char *str;
 };
